+++ edison-oauth/src/main/java/de/otto/edison/oauth/OAuthPublicKeyInMemoryRepository.java
@@ -1,29 +1,28 @@
 package de.otto.edison.oauth;
 
-import org.springframework.util.Assert;
-
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.stream.Collectors;
 
 public class OAuthPublicKeyInMemoryRepository implements OAuthPublicKeyRepository {
 
-    private final List<OAuthPublicKey> activePublicKeys = new ArrayList<>();
+    private final List<OAuthPublicKey> activePublicKeys = new CopyOnWriteArrayList<>();
 
     @Override
-    public void refreshPublicKeys(final List<OAuthPublicKey> publicKeys) throws IllegalArgumentException {
-
-        Assert.isTrue(publicKeys
-                .stream()
-                .allMatch(this::isValid), "Invalid public keys retrieved");
-
+    public void refreshPublicKeys(final List<OAuthPublicKey> publicKeys) {
         activePublicKeys.addAll(publicKeys
                 .stream()
+                .filter(this::isValid)
                 .filter(k -> !activePublicKeys.contains(k))
                 .collect(Collectors.toList())
         );
+    }
+
+    @Override
+    public List<OAuthPublicKey> retrieveActivePublicKeys() {
+        return activePublicKeys.stream().filter(this::isValid).collect(Collectors.toList());
     }
 
     private boolean isValid(final OAuthPublicKey publicKey) {
@@ -31,10 +30,5 @@ public class OAuthPublicKeyInMemoryRepository implements OAuthPublicKeyRepositor
 
         return now.isAfter(publicKey.getValidFrom()) &&
                 (Objects.isNull(publicKey.getValidUntil()) || now.isBefore(publicKey.getValidUntil()));
-    }
-
-    @Override
-    public List<OAuthPublicKey> retrieveActivePublicKeys() {
-        return activePublicKeys.stream().filter(this::isValid).collect(Collectors.toList());
     }
 }